submit

+++ c05/ex05/ft_sqrt.c
@@ -6,7 +6,7 @@
 /*   By: ijeon <ijeon@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/03 15:43:01 by ijeon             #+#    #+#             */
-/*   Updated: 2020/12/06 12:04:48 by ijeon            ###   ########.fr       */
+/*   Updated: 2020/12/06 14:03:02 by ijeon            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,6 +22,6 @@ int	ft_sqrt(int nb)
 		num++;
 	}
 	if (num * num == (unsigned int)nb)
-		return (num);
+		return ((int)num);
 	return (0);
 }
